+++ data/i18n/hydrogen_uk.ts
@@ -1395,6 +1395,26 @@ Please set your system&apos;s locale to UTF-8!</source>
         <source>Save song</source>
         <translation>Записати композицію</translation>
     </message>
+    <message>
+        <source>Editing type of instrument</source>
+        <extracomment>Representing a altering of an instrument type in the undo history and * context menu.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Add layer</source>
+        <extracomment>Representing adding an instrument layer in the undo history. Both the * name of the layer and the corresponding instrument will be appended.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Delete layer</source>
+        <extracomment>Representing deleting an instrument layer in the undo history. Both the * name of the layer and the corresponding instrument will be appended.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Edit layer</source>
+        <extracomment>Representing editing an instrument layer in the undo history. Both the * name of the layer and the corresponding instrument will be appended.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>ComponentView</name>