+++ src/core/CoreActionController.cpp
@@ -1981,6 +1981,29 @@ bool CoreActionController::moveInstrument( int nSourceIndex, int nTargetIndex ) 
 	return true;
 }
 
+bool CoreActionController::setInstrumentType( int nInstrumentId,
+											  const DrumkitMap::Type& sType ) {
+	auto pHydrogen = Hydrogen::get_instance();
+	ASSERT_HYDROGEN
+
+	auto pSong = pHydrogen->getSong();
+	if ( pSong == nullptr || pSong->getDrumkit() == nullptr ) {
+		return false;
+	}
+
+	auto pInstrument = pSong->getDrumkit()->getInstruments()->
+		find( nInstrumentId );
+	if ( pInstrument == nullptr ) {
+		ERRORLOG( QString( "Uable to find instrument [%1] for type [%2]" )
+				  .arg( nInstrumentId ).arg( sType ) );
+		return false;
+	}
+
+	pInstrument->setType( sType );
+
+	return true;
+}
+
 bool CoreActionController::locateToColumn( int nPatternGroup ) {
 	auto pHydrogen = Hydrogen::get_instance();
 	ASSERT_HYDROGEN