+++ src/core/CoreActionController.h
@@ -26,6 +26,7 @@
 #include <vector>
 #include <memory>
 
+#include <core/Basics/DrumkitMap.h>
 #include <core/Object.h>
 
 namespace H2Core
@@ -360,6 +361,15 @@ class CoreActionController : public H2Core::Object<CoreActionController> {
 		 * within the instrument list and _not_ the ID of the instrument (which
 		 * stays the same during the move action). */
 		static bool moveInstrument( int nSourceIndex, int nTargetIndex );
+
+		/** Changing the type of an instrument - specified using its ID - of the
+		 * current drumkit.
+		 *
+		 * This one should be used over setDrumkit() since it is able to
+		 * properly handling setting the initial type or removing a type string
+		 * while honoring JACK per-track output ports. */
+		static bool setInstrumentType( int nInstrumentId,
+									   const DrumkitMap::Type& sType );
 
 		/** Relocates transport to the beginning of a particular
 		 * column/Pattern group.