+++ src/gui/src/InstrumentEditor/ComponentView.cpp
@@ -800,10 +800,30 @@ void ComponentView::removeLayerButtonClicked() {
 		return;
 	}
 
+	auto pHydrogenApp = HydrogenApp::get_instance();
+
+	const auto pInstrument = pHydrogenApp->getInstrumentRack()->
+		getInstrumentEditorPanel()->getInstrument();
+	auto pNewInstrument = std::make_shared<Instrument>( pInstrument );
+	auto pNewComponent = pNewInstrument->getComponent(
+		pInstrument->index( m_pComponent ) );
+	if ( pNewComponent == nullptr ) {
+		ERRORLOG( "Hiccup while looking up component" );
+		return;
+	}
+
+	auto pLayer = m_pComponent->getLayer( m_nSelectedLayer );
+	if ( pLayer == nullptr ) {
+		// Nothing to remove
+		return;
+	}
+	const QString sLayerName = pLayer->getSample() != nullptr ?
+		pLayer->getSample()->getFilename() : "nullptr";
+
 	auto pHydrogen = Hydrogen::get_instance();
 	pHydrogen->getAudioEngine()->lock( RIGHT_HERE );
 
-	m_pComponent->setLayer( nullptr, m_nSelectedLayer );
+	pNewComponent->setLayer( nullptr, m_nSelectedLayer );
 
 	pHydrogen->getAudioEngine()->unlock();
 
@@ -828,6 +848,11 @@ void ComponentView::removeLayerButtonClicked() {
 
 	setSelectedLayer( nCount );
 	updateView();
+
+	pHydrogenApp->pushUndoCommand(
+		new SE_replaceInstrumentAction(
+			pNewInstrument, pInstrument,
+			SE_replaceInstrumentAction::Type::DeleteLayer, sLayerName ) );
 }
 
 void ComponentView::loadLayerBtnClicked() {
@@ -888,7 +913,7 @@ void ComponentView::loadLayerBtnClicked() {
 		return;
 	}
 
-	auto pInstrument = pHydrogenApp->getInstrumentRack()->
+	const auto pInstrument = pHydrogenApp->getInstrumentRack()->
 		getInstrumentEditorPanel()->getInstrument();
 	bool bRenameInstrument = false;
 	if ( filename[0] == "true" ){
@@ -896,6 +921,15 @@ void ComponentView::loadLayerBtnClicked() {
 	}
 	QString sNewInstrumentName;
 
+	auto pNewInstrument = std::make_shared<Instrument>( pInstrument );
+	auto pNewComponent = pNewInstrument->getComponent(
+		pInstrument->index( m_pComponent ) );
+	if ( pNewComponent == nullptr ) {
+		ERRORLOG( "Hiccup while looking up component" );
+		return;
+	}
+
+	QStringList newLayersPaths;
 	int nLastInsertedLayer = m_nSelectedLayer;
 	if ( filename.size() > 2 ) {
 		for ( int ii = 2; ii < filename.size(); ++ii ) {
@@ -906,12 +940,13 @@ void ComponentView::loadLayerBtnClicked() {
 			}
 
 			auto pNewSample = Sample::load( filename[ii] );
+			newLayersPaths << filename[ ii ];
 
 			pHydrogen->getAudioEngine()->lock( RIGHT_HERE );
 
 			// If we're using multiple layers, we start inserting the first
 			// layer at nSelectedLayer and the next layer at nSelectedLayer + 1.
-			auto pLayer = m_pComponent->getLayer( nnLayer );
+			auto pLayer = pNewComponent->getLayer( nnLayer );
 			if ( pLayer != nullptr ) {
 				// insert new sample from newInstrument, old sample gets deleted
 				// by setSample
@@ -919,7 +954,7 @@ void ComponentView::loadLayerBtnClicked() {
 			}
 			else {
 				pLayer = std::make_shared<H2Core::InstrumentLayer>( pNewSample );
-				m_pComponent->setLayer( pLayer, nnLayer );
+				pNewComponent->setLayer( pLayer, nnLayer );
 			}
 			nLastInsertedLayer = nnLayer;
 
@@ -946,20 +981,19 @@ void ComponentView::loadLayerBtnClicked() {
 	// The user choose to rename the instrument according to the (last) filename
 	// of the selected sample.
 	if ( bRenameInstrument && ! sNewInstrumentName.isEmpty() ) {
-		auto pNewInstrument = std::make_shared<Instrument>( pInstrument );
 		pNewInstrument->setName( sNewInstrumentName );
-
-		pHydrogenApp->pushUndoCommand(
-			new SE_replaceInstrumentAction(
-				pNewInstrument, pInstrument,
-				SE_replaceInstrumentAction::Type::RenameInstrument,
-				sNewInstrumentName, pInstrument->getName() ) );
 
 		pHydrogenApp->showStatusBarMessage(
 			QString( "%1 [%2] -> [%3]" )
 			.arg( pHydrogenApp->getCommonStrings()->getActionRenameInstrument() )
 			.arg( pInstrument->getName() ).arg( sNewInstrumentName ) );
 	}
+
+	pHydrogenApp->pushUndoCommand(
+		new SE_replaceInstrumentAction(
+			pNewInstrument, pInstrument,
+			SE_replaceInstrumentAction::Type::AddLayer,
+			newLayersPaths.join( " " ) ) );
 }
 
 void ComponentView::setAutoVelocity() {